pyproject.toml
[tool.poetry]
name = "poc-python-selenium"
version = "0.1.0"
description = ""
authors = ["Ivan Ogasawara <ivan.ogasawara@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "poc_python_selenium"}]

[tool.poetry.dependencies]
python = "^3.9"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
pre-commit = "^3.2.2"
makim = "1.6.5"
containers-sugar = "1.4.4"
docker-compose = "^1.29.2"
selenium = "^4.3.0"
chromedriver-binary-auto = "^0.2.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
